requirements.txt
alembic==1.9.2
bcrypt==4.0.1
Bcrypt-Flask==1.0.2
Jinja2==2.11.3
SQLAlchemy==1.3.0
Werkzeug==1.0.1
WTForms==3.0.1
psycopg2
Flask-COMBO-JSONAPI==1.1.0
Flask==1.1.4
Flask-Admin==1.6.1
Flask-Login==0.6.2
Flask-Migrate==4.0.2
Flask-SQLAlchemy==2.3.0
Flask-WTF==1.1.1
Mako==1.2.4
MarkupSafe==2.0.1
dnspython==2.3.0
email-validator==1.3.1
greenlet==2.0.1
idna==3.4
python-dotenv==0.21.1
psycopg2
ComboJSONAPI==1.1.2
apispec==4.7.1
PyYAML==5.3.1
requests
gunicorn==21.2.0